requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
numpy
pandas
scikit-learn
tensorflow-cpu
joblib